Check Data for BRMA Workshop

This function checks that argument df is a suitable data.frame for completing the workshop on Bayesian Regularized Meta-Regression, and checks that pema package dependencies are correctly installed. It suggests how to remedy any failed checks.

Usage

check_workshop_data(df)

Arguments

df

A data.frame.

Value

Invisibly returns a logical TRUE/FALSE, indicating whether all checks have passed.

Examples

check_workshop_data(fukkink_lont)
#> Warning: restarting interrupted promise evaluation
#> Warning: restarting interrupted promise evaluation