Sample from the Prior Distribution

Samples from a prior distribution with parameters defined in prior. The result can be plotted using the plot function.

Usage

sample_prior(
  method = c("hs", "lasso"),
  prior = switch(method, lasso = c(df = 1, scale = 1), hs = c(df = 1, df_global = 1,
    df_slab = 4, scale_global = 1, scale_slab = 2, par_ratio = NULL)),
  iter = 1000
)

Arguments

method

Character string, indicating which prior to sample from. Default: first element of c("hs", "lasso").

prior

Numeric vector, specifying the prior to use. See brma for more details.

iter

A positive integer specifying the number of iterations to sample. Default: 1000

Value

NULL, function is called for its side-effect of plotting to the graphics device.

Examples

sample_prior("lasso", iter = 10)
#> $method
#> [1] "lasso"
#> 
#> $iter
#> [1] 10
#> 
#> $samples
#> Inference for Stan model: lasso_prior.
#> 1 chains, each with iter=10; warmup=0; thin=1; 
#> post-warmup draws per chain=10, total post-warmup draws=10.
#> 
#>                   mean se_mean   sd  2.5%   25%   50%   75% 97.5% n_eff Rhat
#> b                -0.26    0.66 1.71 -2.91 -1.23  0.05  1.18  1.72     7 0.89
#> lasso_inv_lambda  1.80    0.50 1.59  0.36  0.86  0.86  2.49  4.60    10 0.90
#> lp__             -4.04    0.22 0.71 -5.12 -4.57 -3.90 -3.41 -3.32    10 0.99
#> 
#> Samples were drawn using NUTS(diag_e) at Sat Mar 29 16:35:06 2025.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).
#> 
#> attr(,"class")
#> [1] "brma_prior" "list"